+++ TestApplications/src/Notifier/Program.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -16,16 +17,29 @@ namespace NotifierTest
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            if(args.Length > 1)
+            if (args.Length == 1)
             {
-                var title = args.Aggregate("", (current, arg) => current + arg);
-                Application.Run(new Notifier(title, true));
+                if (args[0] == "Delayed")
+                {
+                    Thread.Sleep(3000);
+                }
+                RunNotifierWithoutTimer(args);
+            }
+            if(args.Length >= 1)
+            {
+                RunNotifierWithoutTimer(args);
             }
             else
             {
                 Application.Run(new Notifier());
             }
             
+        }
+
+        static void RunNotifierWithoutTimer(string[] args)
+        {
+            var title = args.Aggregate("", (current, arg) => current + arg);
+            Application.Run(new Notifier(title, true));
         }
     }
 }