+++ src/FlaUILibrary/flaui/util/__init__.py
@@ -2,3 +2,4 @@ from .keyboardinputconverter import KeyboardInputConverter
 from .interfacetype import InterfaceType
 from .treeitems import TreeItems
 from .treeitemsparser import TreeItemsParser
+from .util import *